Move sdrctl testmode cmd out to sdrctl_intf.c
